chore(sdk): move FailOnNondeterminismInterceptor to tests
