+++ CHANGELOG.md
@@ -52,6 +52,7 @@ relevant information.
 ### Breaking Changes :boom:
 * Removed the unused `ActExitValue` type. Use `ActivityError::WillCompleteAsync` to mark an
   activity for asynchronous completion.
+* Removed the test-only `FailOnNondeterminismInterceptor` from the public Rust SDK API.
 * `IncomingError` is now non-exhaustive. Add a wildcard branch to downstream matches over decoded
   failures.
 * Rust SDK environment configuration values (`DataSource`, `ClientConfig`, and related profile,
